refactor(components): abstracts collection search view into shared layout
